+++ src/main/java/be/inniger/problems/Day07.java
@@ -8,19 +8,19 @@ import java.util.List;
 
 public class Day07 {
 
-    public static int partOne(List<Integer> program) {
-        return bothParts(program, List.of(0, 1, 2, 3, 4));
+    public static long partOne(List<Long> program) {
+        return bothParts(program, List.of(0L, 1L, 2L, 3L, 4L));
     }
 
-    public static int partTwo(List<Integer> program) {
-        return bothParts(program, List.of(5, 6, 7, 8, 9));
+    public static long partTwo(List<Long> program) {
+        return bothParts(program, List.of(5L, 6L, 7L, 8L, 9L));
     }
 
-    private static int bothParts(List<Integer> program, List<Integer> phasesElements) {
-        return Util.permutations(phasesElements).stream().mapToInt(phases -> amplify(program, phases)).max().orElseThrow();
+    private static long bothParts(List<Long> program, List<Long> phasesElements) {
+        return Util.permutations(phasesElements).stream().mapToLong(phases -> amplify(program, phases)).max().orElseThrow();
     }
 
-    private static int amplify(List<Integer> program, List<Integer> phases) {
+    private static long amplify(List<Long> program, List<Long> phases) {
         // Create IntCode computers, with the phase as their first input
         var intCodes = phases.stream().map(phase -> phasedIntCode(program, phase)).toList();
 
@@ -31,7 +31,7 @@ public class Day07 {
         }
 
         // Additionally, once, give the value 0 as input to the first computer
-        intCodes.getFirst().input().add(0);
+        intCodes.getFirst().input().add(0L);
 
         // Keep running the computers until the last one has halted
         var status = Status.RUNNING;
@@ -45,7 +45,7 @@ public class Day07 {
         return intCodes.getLast().output().remove();
     }
 
-    private static IntCode phasedIntCode(List<Integer> program, Integer phase) {
+    private static IntCode phasedIntCode(List<Long> program, Long phase) {
         var intCode = new IntCode(program);
         intCode.input().add(phase);
         return intCode;
